fix: resolve all cargo clippy -D warnings across all five workspaces
ingress:
- Remove unused imports (Claims, jsonwebtoken, GrpcHealthCheck, GrpcServerConfig, GrpcServerConfigError, MAX_MESSAGE_BYTES, IngressTlsConfig, MAX_BODY_BYTES, HandlerRegistry, Arc)
- Fix identical if-blocks in health_service::recompute_overall_status
- Fix first().is_none() -> is_empty() in peer_identity
- Elide explicit lifetime 'a in parse_atv and extract_bearer
- Suppress result_large_err on verify_auth (axum::response::Response is intentionally large)
- Add missing doc comments to ServingStatus variants, constants, AxumServerError variants, TonicServerError variants, GrpcServerConfigError variant, TonicGrpcServer builder methods
- Add reflection service constant docs, remove unused Arc and HandlerRegistry imports

egress:
- Remove redundant pub use re-exports from api/traits, api/client, api/config, api/resilience, api/status_codes, api/value_object/grpc
- Gate to_tonic_code/to_wire behind cfg(test); remove Conversions marker structs
- Add #[allow(dead_code)] to unused builder structs/impls (GrpcClientBuilder, GrpcChannelConfigBuilder, GrpcRequestBuilder, TonicGrpcClientBuilder, ResilienceValidator, Processor)
- Remove unused MtlsAuthConfig, AuthzPolicy, MethodAclConfig, Arc imports
- Fix redundant u64 cast in backoff::from_clock
- Add missing docs to oauth Error variants

runtime:
- Remove unused pub use re-exports in composite, json_codec, monitor, runner, traits
- Add #[allow(clippy::module_inception)] to 10 module-inception sites
- Add missing doc comments to RuntimeBuilder methods, ConfigError/RuntimeError variants, Input/Output trait methods, ComponentHealth/RuntimeHealth fields, Runner::run, lifecycle_monitor module
- Fix TracingFormat import in observability.rs (inline in cfg block)

domain:
- Remove unused Handler re-export from api/traits; add Default impl for RequestContextBuilder
- Remove unused Validator re-export and ValidateResult type alias from api/validator
- Add #[allow(dead_code)] to ValidatorDefault, find_handler, format_error
- Add #[allow(clippy::module_inception)] to api/handler/mod.rs

proxy:
- Remove unused NoopValidator re-exports; add #[allow(dead_code)] to Monitor, NoopValidator trait and struct

+++ main/features/retry/src/core/backoff.rs
@@ -107,7 +107,7 @@ impl JitterRng {
     pub(crate) fn from_clock() -> Self {
         let nanos = std::time::SystemTime::now()
             .duration_since(std::time::UNIX_EPOCH)
-            .map(|d| d.subsec_nanos() as u64 ^ (d.as_secs() as u64).rotate_left(17))
+            .map(|d| u64::from(d.subsec_nanos()) ^ d.as_secs().rotate_left(17))
             .unwrap_or(0);
         Self::new(nanos.wrapping_add(0xDEADBEEF))
     }
